+++ src/main.rs
@@ -279,6 +279,8 @@ fn generate_schema() -> anyhow::Result<()> {
         use itertools::Itertools;
         use strum::VariantArray;
 
+        // TODO: Maybe make this case insensitive? See https://stackoverflow.com/a/68639341
+
         let proc_columns = schema.definitions.get_mut("ProcColumn").unwrap();
         match proc_columns {
             schemars::schema::Schema::Object(proc_columns) => {
@@ -292,6 +294,21 @@ fn generate_schema() -> anyhow::Result<()> {
                     .collect();
             }
             _ => anyhow::bail!("missing proc columns definition"),
+        }
+
+        let disk_columns = schema.definitions.get_mut("DiskColumn").unwrap();
+        match disk_columns {
+            schemars::schema::Schema::Object(disk_columns) => {
+                let enums = disk_columns.enum_values.as_mut().unwrap();
+                *enums = widgets::DiskColumn::VARIANTS
+                    .iter()
+                    .flat_map(|var| var.get_schema_names())
+                    .sorted()
+                    .map(|v| serde_json::Value::String(v.to_string()))
+                    .dedup()
+                    .collect();
+            }
+            _ => anyhow::bail!("missing disk columns definition"),
         }
     }