+++ src/options/config/network.rs
@@ -6,7 +6,7 @@ use super::IgnoreList;
 #[derive(Clone, Debug, Default, Deserialize)]
 #[cfg_attr(feature = "generate_schema", derive(schemars::JsonSchema))]
 #[cfg_attr(test, serde(deny_unknown_fields), derive(PartialEq, Eq))]
-pub struct NetworkConfig {
+pub(crate) struct NetworkConfig {
     /// A filter over the network interface names.
-    pub interface_filter: Option<IgnoreList>,
+    pub(crate) interface_filter: Option<IgnoreList>,
 }